requirements.txt
absl-py==1.3.0
alabaster==0.7.13
altgraph==0.17.3
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.11.1
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.2.0
cattrs==23.1.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
decorator==5.1.1
distro==1.8.0
Django==4.1.4
docutils==0.20.1
esbonio==0.16.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
Flask==2.2.2
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
ghostscript==0.7
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
imagesize==1.4.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kaitaistruct==0.10
keras==2.11.0
kiwisolver==1.4.4
ldap3==2.9.1
libclang==14.0.6
lsprotocol==2023.0.0b1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib @ file:///C:/WINDOWS/system32/matplotlib-3.5.2-cp310-cp310-win_amd64.whl
matplotlib-inline==0.1.6
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.19
msgpack==1.0.4
nltk==3.8
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
pdfminer.six==20221105
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.38
protobuf==3.19.6
publicsuffix2==2.20191221
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydivert==2.1.0
pygame==2.5.0
pygls==1.1.1
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyMuPDF==1.22.3
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypdf==3.9.1
PyPDF2==3.0.1
pyperclip==1.8.2
pyspellchecker==0.7.2
python-dateutil==2.8.2
pytz==2022.7
pywin32-ctypes==0.2.2
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.13.5
tabula-py==2.7.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typeguard==3.0.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
urwid==2.1.2
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zstandard==0.19.0
